Added `persistDestination` option to diff trigger creation. Allows the destination table to be persisted beyond cleanup of a trigger, also allows the creation to complete even if the destination table already exists.

+++ packages/common/src/client/triggers/TriggerManagerImpl.ts
@@ -201,6 +201,7 @@ export class TriggerManagerImpl implements TriggerManager {
       columns,
       when,
       hooks,
+      persistDestination = false,
       // Fall back to the provided default if not given on this level
       useStorage = this.defaultConfig.useStorageByDefault
     } = options;
@@ -268,11 +269,13 @@ export class TriggerManagerImpl implements TriggerManager {
      * we need to ensure we can cleanup the created resources.
      * We unfortunately cannot rely on transaction rollback.
      */
-    const cleanup = async () => {
+    const cleanup = async (force?: boolean) => {
       disposeWarningListener();
       return this.db.writeLock(async (tx) => {
         await this.removeTriggers(tx, triggerIds);
-        await tx.execute(/* sql */ `DROP TABLE IF EXISTS ${destination};`);
+        if (!persistDestination || force) {
+          await tx.execute(/* sql */ `DROP TABLE IF EXISTS ${destination};`);
+        }
         await releaseStorageClaim?.();
       });
     };
@@ -281,7 +284,7 @@ export class TriggerManagerImpl implements TriggerManager {
       // Allow user code to execute in this lock context before the trigger is created.
       await hooks?.beforeCreate?.(tx);
       await tx.execute(/* sql */ `
-        CREATE ${tableTriggerTypeClause} TABLE ${destination} (
+        CREATE ${tableTriggerTypeClause} TABLE ${persistDestination ? 'IF NOT EXISTS ' : ''}${destination} (
           operation_id INTEGER PRIMARY KEY AUTOINCREMENT,
           id TEXT,
           operation TEXT,
